pyproject.toml
[project]
name = "economic-signal-engine"
version = "1.0"
description = "Produce rules based signals based on economic data"
authors = [{name = "Batuhan"}]
license = {text= "Apache"}
readme = "README.md"
requires-python = ">=3.6"
dependencies = ["pyyaml"]

[project.urls]
Homepage = "https://github.com/BatuhanOzdemir/economic-signal-engine"

[project.scripts]
ese = "ese.__main__:main"
